requirements.txt
docopt
plaid-python >= 8.0.0
python-dateutil
